Image Carousel & CSS
With this commit, I used Bootstrap 4 to add an image carousel. I also changed the background color and replaced on the home image to a clip art image that was better suited and styled it to match the content. I also added favicons to each of the page's tabs to match its content.