Navigation Style & Header/Footer Color
Revamped the navigation menu by styling hidden buttons, added a border on the bottom, added a matching top-border for the footer, and changed the color of the fonts in the hover effect. Added header and footer color to match the color scheme. Changed the "Restaurant Description" to create colored fonts with black outline. Added a little bit of space before and after the tables in the menu.html file so that the would not overlap on the borders in the header and footer.